+++ examples/flow-react-example/src/Hooks.gen.js
@@ -15,7 +15,9 @@ import type {reactElement as ReasonReact_reactElement} from '../src/shims/ReactS
 
 export type vehicle = {| +name: string |};
 
-export const make: ({| +vehicle: vehicle |}) => ReasonReact_reactElement = function (Arg1: $any) {
-  const result = HooksBS.make({vehicle:[Arg1.vehicle.name]});
+export const $$default: ({| +vehicle: vehicle |}) => ReasonReact_reactElement = function (Arg1: $any) {
+  const result = HooksBS.default({vehicle:[Arg1.vehicle.name]});
   return result
 };
+
+export default $$default;